Combine output and writer logic
